Add external library figures to the memory chart [#718]

+++ app/src/main/java/me/devsaki/hentoid/fragments/queue/QueueFragment.java
@@ -495,7 +495,7 @@ public class QueueFragment extends Fragment implements ItemTouchCallback, Simple
             dlPreparationProgressBar.setVisibility(View.GONE);
         }
 
-        dlPreparationProgressBar.setProgress1((long) event.total - event.done);
+        dlPreparationProgressBar.setProgress1(event.total - event.done);
     }
 
     /**
